EKonsulta/requirements.txt
Flask
fillpdf
pdfrw
pdf2image
mysql-connector-python
python-dotenv
Pillow
waitress
tzdata